[util] Move subset options into subset-options.hh
