+++ util/options.hh
@@ -28,7 +28,6 @@
 #define OPTIONS_HH
 
 #include "hb.hh"
-#include "hb-subset.h"
 
 #include <stdlib.h>
 #include <stddef.h>
@@ -620,45 +619,6 @@ struct format_options_t
   hb_bool_t show_extents = false;
   hb_bool_t show_flags = false;
   hb_bool_t trace = false;
-};
-
-struct subset_options_t
-{
-  subset_options_t ()
-  : input (hb_subset_input_create_or_fail ())
-  {}
-  ~subset_options_t ()
-  {
-    hb_subset_input_destroy (input);
-  }
-
-  void add_options (option_parser_t *parser);
-
-  hb_bool_t* bool_for(hb_subset_flags_t flag)
-  {
-    for (unsigned i = 0; i < sizeof(int) * 8; i++)
-    {
-      if (1u << i == flag)
-        return &flags[i];
-    }
-    return &flags[sizeof(int) * 8 - 1];
-  }
-
-  hb_subset_input_t * get_input ()
-  {
-    hb_subset_flags_t flags_set = HB_SUBSET_FLAGS_DEFAULT;
-    for (unsigned i = 0; i < sizeof(int) * 8; i++)
-    {
-      if (flags[i])
-        flags_set = (hb_subset_flags_t) (flags_set |  (1u << i));
-    }
-    hb_subset_input_set_flags (input, flags_set);
-    return input;
-  }
-
-  unsigned num_iterations = 1;
-  hb_subset_input_t *input = nullptr;
-  hb_bool_t flags[sizeof(int) * 8] = {0};
 };
 
 /* fallback implementation for scalbn()/scalbnf() for pre-2013 MSVC */
