+++ util/subset-options.hh
@@ -24,7 +24,50 @@
  * Google Author(s): Garret Rieger
  */
 
+#ifndef SUBSET_OPTIONS_HH
+#define SUBSET_OPTIONS_HH
+
 #include "options.hh"
+#include "hb-subset.h"
+
+struct subset_options_t
+{
+  subset_options_t ()
+  : input (hb_subset_input_create_or_fail ())
+  {}
+  ~subset_options_t ()
+  {
+    hb_subset_input_destroy (input);
+  }
+
+  void add_options (option_parser_t *parser);
+
+  hb_bool_t* bool_for(hb_subset_flags_t flag)
+  {
+    for (unsigned i = 0; i < sizeof(int) * 8; i++)
+    {
+      if (1u << i == flag)
+        return &flags[i];
+    }
+    return &flags[sizeof(int) * 8 - 1];
+  }
+
+  hb_subset_input_t * get_input ()
+  {
+    hb_subset_flags_t flags_set = HB_SUBSET_FLAGS_DEFAULT;
+    for (unsigned i = 0; i < sizeof(int) * 8; i++)
+    {
+      if (flags[i])
+        flags_set = (hb_subset_flags_t) (flags_set |  (1u << i));
+    }
+    hb_subset_input_set_flags (input, flags_set);
+    return input;
+  }
+
+  unsigned num_iterations = 1;
+  hb_subset_input_t *input = nullptr;
+  hb_bool_t flags[sizeof(int) * 8] = {0};
+};
 
 
 static gboolean
@@ -323,3 +366,5 @@ subset_options_t::add_options (option_parser_t *parser)
 	 "Options subsetting",
 	 this);
 }
+
+#endif
